Refactor to nested reducers
A nested reducer was created in order to handle the actions from the top bar (settings). This will simplify the reducers design pattern and make it easier to read.
Actions were also moved to their own sub-folder.
