+++ src/services/utils.js
@@ -1,0 +1,3 @@
+export default function capitalizeFirstLetter(string) {
+  return string.charAt(0).toUpperCase() + string.slice(1);
+}
